refactor: more intuitively model program structure
Also makes clippy happy on a couple checks.
